requirements.txt
requests~=2.28.1
whois~=0.9.19
Flask~=2.2.2
bs4~=0.0.1
beautifulsoup4~=4.11.1
Pillow~=9.4.0
selenium~=4.5.0